support skip cleanup

+++ examples/rrsa/e2e-test/e2e.sh
@@ -77,8 +77,10 @@ function assume_role() {
 function cleanup() {
   set +e
   bar_tip "cleanup"
-  aliyun ram DeleteRole --RoleName ${ROLE_NAME}
-  rm ${KUBECONFIG_PATH}
+  if ! echo "${SKIP}" |grep cleanup; then
+    aliyun ram DeleteRole --RoleName ${ROLE_NAME}
+    rm ${KUBECONFIG_PATH}
+  fi
   set -e
 }
